fix(src) : 코드 모듈화
모듈화 요구사항을 지키기 위하여
index.js에 있던 코드를 분리함.
